Issue #5439: share CLI helper implementations
